use subpixels instead

+++ src/geometry/circle.hpp
@@ -4,8 +4,8 @@
 
 class Circle : public GameObject {
   public:
-    Circle(int x, int y, int radius, Mask mask) : GameObject(Objects::Circle, x, y, mask), radius(radius) {}
-    int radius;
+    Circle(float x, float y, float radius, Mask mask) : GameObject(Objects::Circle, x, y, mask), radius(radius) {}
+    float radius;
 
     bool collides(GameObject *object) override { return false; }
 };
