+++ src/physics/physics_box.hpp
@@ -6,7 +6,7 @@ class PhysicsCircle;
 
 class PhysicsBox : public Box {
   public:
-    PhysicsBox(int x, int y, int width, int height, Mask mask) : Box(x, y, width, height, mask) {
+    PhysicsBox(float x, float y, float width, float height, Mask mask) : Box(x, y, width, height, mask) {
         Box::identifier = Objects::PhysicsBox;
     };
     bool collides(GameObject *object) override;
